all apis test over

+++ model/model.go
@@ -3,6 +3,8 @@ package model
 import (
 	//"SC/util"
 	"errors"
+	"log"
+	"strconv"
 	"time"
 
 	"github.com/dgrijalva/jwt-go"
@@ -57,6 +59,7 @@ func UpdateUserInfo(user User) error {
 	return result.Error
 }
 
+/*
 func GetAchievement(id string) []string {
 	var achievements []Achievement
 	DB.Where("student_id = ?", id).Find(&achievements)
@@ -65,7 +68,7 @@ func GetAchievement(id string) []string {
 		achs[i] = achievements[i].Achievement
 	}
 	return achs
-}
+}*/
 
 //-----------------------------------------------
 //punch:
@@ -74,21 +77,27 @@ func GetPunchAndNumber(id string) Punchs {
 	var punchs []UsersPunch
 	DB.Where("student_id = ?", id).Find(&punchs)
 	var punchs2 []Punch
+	var Punch Punch
 	for i := 0; i < len(punchs); i++ {
-		punchs2[i].Title = punchs[i].Title
-		punchs2[i].Number = punchs[i].Number
+		Punch.Title = punchs[i].Title
+		Punch.Number = punchs[i].Number
+		punchs2 = append(punchs2, Punch)
 	}
 	Punchs := Punchs{Punchs: punchs2}
 	return Punchs
 
 }
 
-func GetPunchs(Type string) Punchs2 {
+func GetPunchs(TypeID string) Punchs2 {
+	Type := Type(TypeID)
 	var punchs []PunchContent
 	DB.Where("type = ?", Type).Find(&punchs)
 	var punchs2 []Punch2
+	var Punch Punch2
 	for i := 0; i < len(punchs); i++ {
-		punchs2[i].Title = punchs[i].Title
+		Punch.Title = punchs[i].Title
+		Punch.ID = punchs[i].ID
+		punchs2 = append(punchs2, Punch)
 	}
 	Punchs := Punchs2{Punchs: punchs2}
 	return Punchs
@@ -98,29 +107,33 @@ func GetMyPunch(id string) Punchs {
 	var punchs []UsersPunch
 	DB.Where("student_id = ?", id).Find(&punchs)
 	var punchs2 []Punch
+	var Punch Punch
 	for i := 0; i < len(punchs); i++ {
-		punchs2[i].Title = punchs[i].Title
-		punchs2[i].Number = punchs[i].Number
+		Punch.Title = punchs[i].Title
+		Punch.Number = punchs[i].Number
+		punchs2 = append(punchs2, Punch)
 	}
 	Punchs := Punchs{Punchs: punchs2}
 	return Punchs
 }
 
-func TodayPunch(StudentId string, title string) Choice {
+func TodayPunch(StudentId string, TitleID int) Choice {
+	var Punch PunchContent
+	DB.Where("id = ?", TitleID).First(&Punch)
 	var punch PunchHistory
 	today := time.Now().YearDay()
-	result := DB.Where("student_id = ? AND title = ? AND day = ?", StudentId, title, today).First(&punch)
-
+	result := DB.Where("student_id = ? AND title = ? AND day = ?", StudentId, Punch.Title, today).First(&punch)
 	var choice bool
 	if result.Error != nil {
 		choice = false
+	} else {
+		choice = true
 	}
-	choice = true
 	Choice := Choice{Choice: choice}
 	return Choice
 }
 
-func CompletePunch(id string, title string) error {
+func CompletePunch(id string, title string, gold int) error {
 	var punch PunchHistory
 	//result:=DB.Model(&punch).Where("id = ? AND title = ?",id ,title).Update("choice", "hello")
 	punch.Title = title
@@ -129,41 +142,81 @@ func CompletePunch(id string, title string) error {
 	punch.Month = int(time.Now().Month())
 	//punch.Week=time.Now().ISOWeek()
 	punch.Day = time.Now().YearDay()
-	result := DB.Create(&punch)
+	if result := DB.Create(&punch); result.Error != nil {
+		return result.Error
+	}
+	//log.Printf("%v+++++++++++\n", punch)
+
+	//修改用户金币
+	var user User
+	DB.Where("student_id = ? ", id).First(&user)
+	DB.Model(&user).Update("gold", gold+user.Gold)
+	// var u User
+	// u = user
+	// u.Gold += gold
+	// log.Println(user)
+	// log.Println(u)
+	// if result := DB.Model(&u).Update(user); result.Error != nil {
+	// 	return result.Error
+	// }
+	//log.Println(u)
+	s := strconv.Itoa(gold)
+	//创建金币历史
+	history := GoldHistory{
+		StudentID:      id,
+		Time:           time.Now(),
+		ChangeNumber:   gold,
+		ResidualNumber: user.Gold,
+		Reason:         "完成打卡+" + s + "金币",
+	}
+	result := DB.Create(&history)
 	return result.Error
 }
 
-func DeletePunch(id string, title string) error {
+func DeletePunch(id string, title string) (string, error) {
 	var u UsersPunch
-	DB.Where("student_id = ? AND title = ? ", id, title).First(&u)
+	if result := DB.Where("student_id = ? AND title = ? ", id, title).First(&u); result.Error != nil {
+		return "用户未选择该标签", nil
+	}
 	result := DB.Delete(&u)
-	return result.Error
+	return "", result.Error
 }
 
 //-----------------------------------------------
 //default:
 
-func List(Type string) []UserAndNumber {
-	type Result struct {
-		StudentID string
-	}
-	var results []Result
+func List(Type string) ([]UserAndNumber, string) {
+	var PunchHistory []PunchHistory
 	if Type == "month" {
-		DB.Table("punch_histories").Select("student_id").Where("month = ?", int(time.Now().Month())).Scan(&results)
+		DB.Where("month = ?", int(time.Now().Month())).Find(&PunchHistory)
+		//DB.Table("punch_histories").Select("student_id").Where("month = ?", int(time.Now().Month())).Scan(&PunchHistory)
 	}
 	if Type == "week" {
-		DB.Table("punch_histories").Select("student_id").Where("day >= ?", int(time.Now().YearDay())-7).Scan(&results)
+		//result := DB.Where("day >= ?", int(time.Now().YearDay())-7).Find(&PunchHistory)
+		DB.Table("punch_histories").Select("student_id").Where("day >= ?", int(time.Now().YearDay())-7).Scan(&PunchHistory)
+		// if result.Error != nil {
+		// 	log.Println(result.Error)
+		// } else {
+		// 	log.Println("hhhhh")
+		// }
 	}
+	log.Printf("%v--------", Type)
+	log.Printf("%v========", PunchHistory)
 	var s []string
-	for i := 0; i < len(results); i++ {
-		s[i] = results[i].StudentID
+	for i := 0; i < len(PunchHistory); i++ {
+		s = append(s, PunchHistory[i].StudentID)
+	}
+	log.Printf("%v+++++++++\n", len(s))
+	if len(s) == 0 {
+		var a []UserAndNumber
+		return a, "未检索到该时间段的打卡信息"
 	}
 	return GetOrder(s)
 	//var punch PunchHistory
 	//DB.Where("student_id = ? AND month = ? ", int(time.Now().Month())).Find(&punch)
 
 }
-func GetOrder(s []string) []UserAndNumber {
+func GetOrder(s []string) ([]UserAndNumber, string) {
 	var Numbers []UserAndNumber
 	Number := UserAndNumber{StudentId: s[0], Number: 1}
 	Numbers = append(Numbers, Number)
@@ -189,7 +242,7 @@ func GetOrder(s []string) []UserAndNumber {
 		}
 		Numbers[i], Numbers[max] = Numbers[max], Numbers[i]
 	}
-	return Numbers
+	return Numbers, ""
 }
 
 func GetListPrice() ListPrices {
@@ -206,41 +259,57 @@ func GetBackdropPrice() Backdrops {
 	return Backdrops
 }
 
-func ChangeRanking(id string, title string) error {
+func ChangeRanking(id string, ranking string) (error, string) {
 	var listPrice ListPrice
-	DB.Where("title = ? ", title).First(&listPrice)
+	DB.Where("ranking = ? ", ranking).First(&listPrice)
 	var user User
 	DB.Where("student_id = ? ", id).First(&user)
 	if user.Gold < listPrice.Price {
-		return errors.New("金币不足")
+		return nil, "金币不足"
 	}
-	var u User
-	u = user
-	u.Gold -= listPrice.Price
-	result := DB.Model(&u).Update(user)
-	return result.Error
+
+	//修改用户金币
+	DB.Model(&user).Update("gold", user.Gold-listPrice.Price)
+
+	//创建金币历史
+	price := listPrice.Price
+	history := GoldHistory{
+		StudentID:      id,
+		Time:           time.Now(),
+		ChangeNumber:   -price,
+		ResidualNumber: user.Gold,
+		Reason:         "兑换排名：" + ranking,
+	}
+	result := DB.Create(&history)
+	return result.Error, ""
 }
 
-func ChangeBackdrop(id string, BackdropId int) error {
+func ChangeBackdrop(id string, BackdropID int) (error, string) {
 	var backdrop Backdrop
-	DB.Where("backdrop_id = ? ", BackdropId).First(&backdrop)
+	DB.Where("backdrop_id = ? ", BackdropID).First(&backdrop)
 	var user User
 	DB.Where("student_id = ? ", id).First(&user)
 	if user.Gold < backdrop.Price {
-		return errors.New("金币不足")
+		return nil, "金币不足"
 	}
-	var u User
-	u = user
-	u.Gold -= backdrop.Price
-	result := DB.Model(&u).Update(user)
-	if result.Error != nil {
-		return result.Error
+	//修改用户金币
+	DB.Model(&user).Update("gold", user.Gold-backdrop.Price)
+	//创建金币历史
+	s := strconv.Itoa(backdrop.BackdropID)
+	history := GoldHistory{
+		StudentID:      id,
+		Time:           time.Now(),
+		ChangeNumber:   -backdrop.Price,
+		ResidualNumber: user.Gold,
+		Reason:         "兑换背景 " + s,
 	}
+	//log.Printf("%v+++++++++\n---%v\n", BackdropID, backdrop.Price)
+	DB.Create(&history)
 	var usersBackdrop UsersBackdrop
-	usersBackdrop.BackdropID = BackdropId
+	usersBackdrop.BackdropID = BackdropID
 	usersBackdrop.StudentID = id
-	result = DB.Create(&usersBackdrop)
-	return result.Error
+	result := DB.Create(&usersBackdrop)
+	return result.Error, ""
 }
 
 func GetBackdrop(id string) Backdrops {
@@ -248,4 +317,31 @@ func GetBackdrop(id string) Backdrops {
 	DB.Table("users_backdrops").Where("student_id = ? ", id).First(&backdrops)
 	Backdrops := Backdrops{Backdrop: backdrops}
 	return Backdrops
+}
+
+func GetGoldHistory(id string) []GoldHistory {
+	var histories []GoldHistory
+	DB.Where("student_id = ? ", id).Find(&histories)
+	return histories
+}
+
+func CreatePunch(id string, title string) (error, string) {
+	var punch UsersPunch
+	punch.StudentID = id
+	punch.Title = title
+	if result := DB.Where("student_id = ? AND title = ？", id, title).First(&punch); result.Error == nil {
+		return nil, "该标签已选择"
+	}
+	result := DB.Create(&punch)
+	return result.Error, ""
+}
+
+func Type(id string) string {
+	if id == "1" {
+		return "健康"
+	} else if id == "2" {
+		return "运动"
+	} else {
+		return "学习"
+	}
 }
